% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydoublefactorial.R
\name{mydoublefactorial}
\alias{mydoublefactorial}
\title{Double Factorial Function}
\usage{
mydoublefactorial(n)
}
\arguments{
\item{n}{A non-negative integer. The number for which the double factorial is to be computed.}
}
\value{
An integer representing the double factorial of the input n.
}
\description{
This function computes the double factorial (denoted as n!! of a given non-negative integer n).
The double factorial is the product of all integers from n to 1 that have the same parity (odd or even) as n.
}
\examples{
mydoublefactorial(5)
# [1] 15

mydoublefactorial(6)
# [1] 48

}
